backend/app/api/v1/collaboration/collab_analytics_dashboard.py
"""
CollabAnalyticsDashboard : Tableau de bord analytics collaboration
- KPIs, heatmaps, scoring IA, historique, export
- Sécurité : audit, logs, RGPD
- Intégration scalable (FastAPI, MongoDB, Redis)

Auteur : Lead Dev, Data Engineer, ML Engineer
"""

from typing import Dict, Any
import numpy as np

class CollabAnalyticsDashboard:
    """
    Fournit des métriques et visualisations analytiques sur la collaboration.
    """
    def __init__(self):
        pass

    def get_kpis(self, ws_id: str) -> Dict[str, Any]:
        # Mock KPIs : à remplacer par calculs réels
        return {
            "nb_collaborateurs": int(np.random.randint(2, 10)),
            "nb_actions": int(np.random.randint(10, 100)),
            "score_ia": float(np.random.uniform(0.5, 1.0)),
            "engagement": float(np.random.uniform(0.5, 1.0))
        }

    def get_heatmap(self, ws_id: str) -> Dict[str, Any]:
        # Mock heatmap : à remplacer par vraie data
        return {f"jour_{i}": float(np.random.uniform(0, 1)) for i in range(7)}

    def get_history(self, ws_id: str) -> Any:
        # À connecter à SharedWorkspaces/VersionControl
        return [
            {"action": "add_member", "user": "user123", "ts": 1720610000},
            {"action": "edit", "user": "user456", "ts": 1720611000}
        ]

# Exemple d’utilisation :
# dash = CollabAnalyticsDashboard()
# print(dash.get_kpis("ws1")
# print(dash.get_heatmap("ws1")
# print(dash.get_history("ws1")
